move mod select into edit route
